% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{attributeAssign}
\alias{attributeAssign}
\title{assign certain attibutes to an object}
\usage{
attributeAssign(object, new_attributes)
}
\arguments{
\item{object}{The object to which the attributes are assigned}

\item{new_attributes}{a list of attributes}
}
\value{
the object with the assigned attributes
}
\description{
assign certain attibutes to an object
}
