% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Calculate the AUC under the curve}
\usage{
auc(value, time)
}
\arguments{
\item{value}{The PK measurements on original scale}

\item{time}{The time variable}
}
\value{
Returns the calculated AUC
}
\description{
Calculate the AUC under the curve
}
\examples{
2+2
}
