% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{aucFit}
\alias{aucFit}
\title{Area under the curve based on a tail fit}
\usage{
aucFit(fit, from, to = Inf)
}
\arguments{
\item{fit}{A tail fit based on either observed data or censored data}

\item{from}{The time from which the area is calculated}

\item{to}{The end of an interval where the area is calculated}
}
\value{
The fitted area
}
\description{
Area under the curve based on a tail fit
}
\examples{
2+2
}
