% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{aumc}
\alias{aumc}
\title{Calculate the AUC under the moment curve}
\usage{
aumc(value, time)
}
\arguments{
\item{value}{The PK measurements on original scale}

\item{time}{The time variable}
}
\value{
Returns the calculated Area under the curve
}
\description{
Calculate the AUC under the moment curve
}
\examples{
2+2
}
