% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{aumcFit}
\alias{aumcFit}
\title{Area under the moment curve based on a tail fit}
\usage{
aumcFit(fit, from, to = Inf)
}
\arguments{
\item{fit}{A tail fit based on either observed data or censored data}

\item{from}{The start of the interval}

\item{to}{The end the interval, deafults to Inf}
}
\value{
The fitted area under the moment curve
}
\description{
Area under the moment curve based on a tail fit
}
\examples{
2+2
}
