% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailfit.R
\name{bestTailFit}
\alias{bestTailFit}
\title{Get the points used for best tail fit}
\usage{
bestTailFit(value, time, min_points = 3, tol = 1e-04)
}
\arguments{
\item{value}{The PK measurements on original scale}

\item{time}{The time variable}

\item{min_points}{The minimum number of used points}

\item{tol}{The tolerence for best fit. The fit based on most points that
gives an adjusted R2 above the highest observed R2 minus the tolerence is
used}
}
\value{
Returns a list with the number of points used, R-squared, adjusted
  R-squared, and the used points
}
\description{
Get the points used for best tail fit
}
\examples{
# Establish the data
atpt1n <- c(5760, 7200, 8640, 10080, 14400, 20160, 30360, 40320, 50400)
aval   <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, NaN)
lower  <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 0)
upper  <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 3000)

# Fit the tail
bestTailFit(aval[!is.na(aval)], atpt1n[!is.na(aval)])

# For censored data the convention in SAS has been to use the lower limits for the value
best_tail <- bestTailFit(lower, atpt1n)

best_tail

}
