% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailfit.R
\name{bestTailFitCens}
\alias{bestTailFitCens}
\title{Title}
\usage{
bestTailFitCens(lower, upper, time, min_points = 3, tol = 1e-04)
}
\arguments{
\item{lower}{The lower limit of the PK measurements on original scale}

\item{upper}{The upper limit of the PK measurements on original scale}

\item{time}{The time variable}

\item{min_points}{The minimum number of used points}

\item{tol}{The tolerence for best fit. The fit based on most points that}
}
\value{
A list with the number of points used, R-squared, adjusted
  R-squared, and the used points Any censored observation occuring after first
# value used in the tail fit is also given in the list of points
}
\description{
Title
}
\examples{
require(ggplot2)
# Establish the data
atpt1n <- c(5760, 7200, 8640, 10080, 14400, 20160, 30360, 40320, 50400)
aval   <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, NaN)
lower  <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 0)
upper  <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 3000)

# For censored data the convention in SAS has been to use the lower
# limits for the concentration. Any censored observation occuring after first
# value used in the tail fit is also given in the list of points
best_tail <- bestTailFitCens(lower, upper, atpt1n)

best_tail

p <- ggplot(data = best_tail$all_r2, aes(n_points, r2_adj)) + geom_point() +
       geom_hline(yintercept = best_tail$tolerance) +
       labs(y = "Adjusted R-squared", x = "Number of points")
p + geom_point(aes(6, 0.99), colour = 2)
}
