% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{checkCommentFolders}
\alias{checkCommentFolders}
\title{check for comment files}
\usage{
checkCommentFolders(comment_setup)
}
\arguments{
\item{comment_setup}{An object from \code{\link{setUpCommentFolders}}}
}
\value{
List with the files in the various deictories
}
\description{
check for comment files
}
\examples{
 \dontrun{
comment_setup <- initiateCommentData()
checkCommentFolders()
}
}
