% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{clockTime}
\alias{clockTime}
\title{Get clock time for a certain time-zone}
\usage{
clockTime(tzone = "UTC")
}
\arguments{
\item{tzone}{The time zone for the time}
}
\value{
The clock time at the specified time zone
}
\description{
Get clock time for a certain time-zone
}
\examples{
clockTime()
}
