% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_rw_plotter.R
\name{create_comb_plot}
\alias{create_comb_plot}
\title{Create a combined plot with both original scale and log scale plots}
\usage{
create_comb_plot(profile_data, best_fit = FALSE, best_fit_n = 3,
  comments = NULL)
}
\arguments{
\item{profile_data}{The data used}

\item{best_fit}{should the best fit be found automatically}

\item{best_fit_n}{The minimum number of points to use}

\item{comments}{data set with comments, created using createPlotComment}
}
\value{
The combined plot on both original and log scale
}
\description{
Create a combined plot with both original scale and log scale plots
}
