% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_rw_plotter.R
\name{createPlotComment}
\alias{createPlotComment}
\title{Convert comment data for the pk plot}
\usage{
createPlotComment(data, comments)
}
\arguments{
\item{data}{The PK data}

\item{comments}{The complete comment data}
}
\value{
A data set formatted for use in the
}
\description{
Convert comment data for the pk plot
}
