% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_rw_plotter.R
\name{createTableComment}
\alias{createTableComment}
\title{Convert comment data for the pk table}
\usage{
createTableComment(data, comments)
}
\arguments{
\item{data}{The PK data}

\item{comments}{The complete comment data}
}
\value{
A data set formatted for use in the
}
\description{
Convert comment data for the pk table
}
