% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_rw_plotter.R
\name{createTablePDComment}
\alias{createTablePDComment}
\title{Convert comment data for the pd table}
\usage{
createTablePDComment(comments)
}
\arguments{
\item{comments}{The complete comment data}
}
\value{
A data set formatted for use in the
}
\description{
Convert comment data for the pd table
}
