% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{deleteComment}
\alias{deleteComment}
\title{Delete a comment}
\usage{
deleteComment(comment_data, comment_id, user)
}
\arguments{
\item{comment_data}{An object initialised by \code{\link{initCommentData}}.}

\item{comment_id}{The id of the comment to change.}

\item{user}{The initials of the user that deletes the comment.}
}
\value{
A tibble with the updated comment data
}
\description{
Delete a comment
}
