% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpostgrescon.R
\name{getPostgresCon}
\alias{getPostgresCon}
\title{open connection to database}
\usage{
getPostgresCon(dw)
}
\arguments{
\item{dw}{A list with connection details}
}
\value{
A database connection
}
\description{
open connection to database
}
\examples{
\dontrun{
# open connection in accordance with details stored in datawarehouse
dw <- config::get("datawarehouse")
con <- getPostgresCon(dw)
}
}
