% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minutes.R
\name{headerMinutes}
\alias{headerMinutes}
\title{Create header for minutes}
\usage{
headerMinutes(x, project, trial, participants = "", excused = "",
  ref = "", title = c(paste0("NN", project, "-", trial),
  "Profile review meeting"), date = format(Sys.Date(), "\%d \%b \%Y"))
}
\arguments{
\item{x}{The word minutes template}

\item{project}{The project number}

\item{trial}{The Trial number}

\item{participants}{The participants}

\item{excused}{Those excused}

\item{ref}{The minute keeper}

\item{title}{Titles for the header}

\item{date}{The date for the minutes}
}
\value{
An object of class rdocx
}
\description{
Create header for minutes
}
