% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{initCommentData}
\alias{initCommentData}
\title{Initialise the comment data}
\usage{
initCommentData(path = getwd(), users = "", profiles = "",
  update_flag_data = FALSE, save_flag_data = TRUE, save = TRUE,
  rds = TRUE, folder_name = "comment_data")
}
\arguments{
\item{path}{The path where the comment setup should be initialised}

\item{users}{The user that are expected to review}

\item{profiles}{Names of the profiles under review.}

\item{update_flag_data}{Should the \code{flag_data} be updated in the data}

\item{save_flag_data}{Should the flag_data be updated on the disk}

\item{save}{Should the elements of the comments be saved}

\item{rds}{Should the files be saved as rds. The rds files are much faster to
work with and lower risk of errors. However, saving the files as csv has
the benefit of being readable outside R.}

\item{folder_name}{Name of the folder to store the review status and comments}
}
\value{
The comment data
}
\description{
Initialise the comment data
}
