% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{linearIntPol}
\alias{linearIntPol}
\title{Do linear interpolation between points}
\usage{
linearIntPol(value, time)
}
\arguments{
\item{value}{The PK measurements on original scale}

\item{time}{The time variable}
}
\value{
The variable given to value with missing points filled in by linear interpolation
}
\description{
Do linear interpolation between points
}
\examples{
2+2
}
