% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minutes.R
\name{minutesPD}
\alias{minutesPD}
\title{Create minutes for a PD meeting}
\usage{
minutesPD(project, trial, participants = "", excused = "", ref = "",
  comments, overall_desc = NULL)
}
\arguments{
\item{project}{The project number}

\item{trial}{The Trial number}

\item{participants}{The participants}

\item{excused}{Those excused}

\item{ref}{The minute keeper}

\item{comments}{The minutes}

\item{overall_desc}{over all description for minutes}
}
\value{
An object of class rdocx
}
\description{
Create minutes for a PD meeting
}
\examples{
\dontrun{

# Get example comments
data(pd_comments)

# Create the minutes
doc <- minutes(project = "1436", trial = "4225", 
               participants = c("SFFL", "LSBD"),
               excused = c(""),
               ref = "",
               comments = pd_comments,
               overall_desc = "Overall the PD profiles looks good")

# Write to a temp file
file <- tempfile(fileext = ".docx") 
print(doc, target = file)

# Open the file
system(paste("open", file))

# Delete the temporary file again
unlink(file)

}

}
