% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pk_comments-data}
\alias{pk_comments-data}
\alias{pk_comments}
\title{PK example comments}
\description{
An example of pk comments
}
\examples{
data(pk_comments)
print(pk_comments)
}
\keyword{data}
