% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpd_environment.R
\docType{data}
\name{pkpd_environment}
\alias{pkpd_environment}
\alias{nnr_environment}
\alias{nnr_env}
\title{The NNPKPD-package enviroment}
\format{An object of class \code{environment} of length 0.}
\usage{
pkpd_environment
}
\description{
An environment for various NNPKPD package objects for session caching and
package parameters. Used internally by the NNPKPD package.
}
\details{
This environment is used by the shiny application \code{\link{runPKPDreview}}.

Using an environment for this is useful as we do not
accidentially create copies of the environment.
}
\keyword{internal}
