% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkPlot.R
\name{pkPlot}
\alias{pkPlot}
\title{Inititiate a PK plot}
\usage{
pkPlot(data, time, value, log = FALSE, ...)
}
\arguments{
\item{data}{\code{data.frame} or tibble containing the data}

\item{time}{\code{character} with the name of the time column}

\item{value}{\code{character} with the name of the value column}

\item{log}{\code{logical} indicating if the dat should be log-transformed}

\item{...}{Additional parameters passed to \code{aes_string}}
}
\value{
Initiates the pk plot and returns an object of class ggplot
}
\description{
Inititiate a PK plot
}
\examples{
require(ggplot2)

adpc_sub <-
  data.frame(
    conc = c(NA, 13290, 48790, 62740, 81510, 89140, 69410, 71490,
             66720, 63170, 65890, 69520, 66660, 76340, 63220, 69020, 59630, 49690,
             50130, 45410, 29980, 19650, 7488, 3332, NA, NA),
    atpt1n  =  c(-15, 60, 180, 360, 720, 900, 1080, 1260, 1440, 1620, 1800, 2160, 2520, 2880,
                 3600, 4440, 5760, 7200, 8640, 10080, 14400, 20160, 30360, 40320, 50400,
                 61920),
    lower_conc =  c(0, 13290, 48790, 62740, 81510, 89140, 69410, 71490, 66720, 63170, 65890,
                    69520, 66660, 76340, 63220, 69020, 59630, 49690, 50130, 45410, 29980, 19650,
                    7488, 3332, 0, 0),
    upper_conc = c(3000, 13290, 48790, 62740, 81510, 89140, 69410, 71490,
                   66720, 63170, 65890, 69520, 66660, 76340, 63220, 69020, 59630, 49690,
                   50130, 45410, 29980, 19650, 7488, 3332, 3000, 3000),
    tailfl = c("", "", "", "", "", "", "", "", "", "", "", "", "", "", "", "",
               "Y", "Y", "Y", "Y", "Y", "Y", "Y", "Y", "Y", ""))
# Create a tail fit
best_tail <- bestTailFitCens(lower = adpc_sub$lower_conc[adpc_sub$tailfl == "Y"],
                             upper = adpc_sub$upper_conc[adpc_sub$tailfl == "Y"],
                             time  = adpc_sub$atpt1n[adpc_sub$tailfl == "Y"])

# Plot the data on original scale
pkPlot(adpc_sub, time = "atpt1n", value = "conc", colour = "tailfl") +
  pkPlot_dot() +
  pkPlot_tailFit(best_tail$best_fit, colour = "#3F9C35") +
  pkPlot_lloq(data = adpc_sub,  time = "atpt1n",
              lower = "lower_conc", upper = "upper_conc")


# Plot the data on log scale
pkPlot(adpc_sub, time = "atpt1n", value = "conc", log = TRUE, colour = "tailfl") +
  pkPlot_dot() +
  pkPlot_tailFit(best_tail$best_fit, log = TRUE, from = 96, colour = "#3F9C35") +
  pkPlot_lloq(data = adpc_sub,  time = "atpt1n", 
              lower = "lower_conc", upper = "upper_conc", log = TRUE)

}
