% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailfit.R
\name{plot.bestTail}
\alias{plot.bestTail}
\title{Plot the adjusted R-square for the different tail fits}
\usage{
\method{plot}{bestTail}(x, ...)
}
\arguments{
\item{x}{An objects returned from \code{bestTailFitCens} of \code{bestTailFit}}

\item{...}{Included to be in compliance with the plot generic}
}
\value{
An ggplot objects
}
\description{
Plot the adjusted R-square for the different tail fits
}
\examples{
# Establish the data
atpt1n <- c(5760, 7200, 8640, 10080, 14400, 20160, 30360, 40320, 50400)
aval   <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, NaN)
lower  <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 0)
upper  <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 3000)

# For censored data the convention in SAS has been to use the lower
# limits for the concentration. Any censored observation occuring after first
# value used in the tail fit is also given in the list of points
best_tail <- bestTailFitCens(lower, upper, atpt1n)

plot(best_tail)
}
