% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailfit.R
\name{plotTailFit}
\alias{plotTailFit}
\title{Plot of pk data}
\usage{
plotTailFit(data = NULL, fit, time, value, lower = NULL,
  upper = NULL, colour = NULL, ..., log_trans = TRUE,
  plot_tail = TRUE, connect_the_dots = TRUE, fitted_points = NULL,
  highlight_points = fitted_points, fit_point_adj = FALSE)
}
\arguments{
\item{data}{Optinal, data set that includes the data to plot}

\item{fit}{The fitted tail fit}

\item{time}{The time variable to plot}

\item{value}{the value to plot}

\item{lower}{optional, lower limit}

\item{upper}{optional, upper limin}

\item{colour}{colour variable for ggplot}

\item{...}{Additional arguments passed on to ggplot aes}

\item{log_trans}{Should the y-axis be log-transformed}

\item{plot_tail}{Should the tail be plotted}

\item{connect_the_dots}{should linear interpolation be shown between the dots}

\item{fitted_points}{optional, The timepoints which the tail fit is based on.}

\item{highlight_points}{highlighted point}

\item{fit_point_adj}{should the tail fit be drawn from the first point before
to the first point after the points given in fitted points}
}
\value{
A ggplot with data and tail fit
}
\description{
Plot of pk data
}
\examples{
require(tibble)
adpc_sub <-
   tibble(
       conc = c(NA, 13290, 48790, 62740, 81510, 89140, 69410, 71490,
                 66720, 63170, 65890, 69520, 66660, 76340, 63220, 69020, 59630, 49690,
                 50130, 45410, 29980, 19650, 7488, 3332, NA, NA),
       atpt1n  =  c(-15, 60, 180, 360, 720, 900, 1080, 1260, 1440, 1620, 1800, 2160, 2520, 2880,
                3600, 4440, 5760, 7200, 8640, 10080, 14400, 20160, 30360, 40320, 50400,
                61920),
       lower_conc =  c(0, 13290, 48790, 62740, 81510, 89140, 69410, 71490, 66720, 63170, 65890,
                69520, 66660, 76340, 63220, 69020, 59630, 49690, 50130, 45410, 29980, 19650,
                7488, 3332, 0, 0),
       upper_conc = c(3000, 13290, 48790, 62740, 81510, 89140, 69410, 71490,
                 66720, 63170, 65890, 69520, 66660, 76340, 63220, 69020, 59630, 49690,
                 50130, 45410, 29980, 19650, 7488, 3332, 3000, 3000),
       tailfl= c("", "", "", "", "", "", "", "", "", "", "", "", "", "", "", "",
                 "Y", "Y", "Y", "Y", "Y", "Y", "Y", "Y", "Y", "")          )


# Example using bestTailFit
best_tail <- bestTailFitCens(adpc_sub$lower_conc[adpc_sub$tailfl == "Y"],
                             adpc_sub$upper_conc[adpc_sub$tailfl == "Y"],
                             adpc_sub$atpt1n[adpc_sub$tailfl == "Y"])

plotTailFit(value = conc, time = atpt1n, data = adpc_sub[!is.na(adpc_sub$conc), ],
            fit = best_tail$best_fit, fitted_points = best_tail$time_points,
            log_trans = FALSE)

plotTailFit(value = conc, time = atpt1n, data = adpc_sub[!is.na(adpc_sub$conc), ],
            fit = best_tail$best_fit, fitted_points = best_tail$time_points,
            log_trans = TRUE)

#plotTailFit(value = conc, time = "atpt1n", lower = lower_conc, upper = upper_conc,
#           data = adpc_sub,
#             fit = best_tail$best_fit, fitted_points = best_tail$time_points,
#             colour = tailfl,
#             log_trans = FALSE)

#plotTailFit(value = conc, time = atpt1n, lower = lower_conc, upper = upper_conc,
#             data = adpc_sub,
#           fit = best_tail$best_fit, fitted_points = best_tail$time_points,
#           colour = tailfl,
#           log_trans = TRUE)
#

#plotTailFit(value = data$conc, time = data$atpt1n, lower = data$lower_conc,
#            upper = data$upper_conc,
#            fit = best_tail$best_fit, fitted_points = best_tail$time_points,
#             colour = data$tailfl,
#             log_trans = TRUE)

}
