% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{predictConc}
\alias{predictConc}
\title{Predic the concentration based on a tail fil}
\usage{
predictConc(fit, time)
}
\arguments{
\item{fit}{A tail fit based on either observed data or censored data}

\item{time}{the time for which the estimation is made.}
}
\value{
The estimated concentration
}
\description{
Predic the concentration based on a tail fil
}
\examples{
2+2
}
