% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{readComDelete}
\alias{readComDelete}
\title{Read the comment deletions}
\usage{
readComDelete(path, files = NULL)
}
\arguments{
\item{path}{path to the deletions}

\item{files}{The files to read, if NULL all files are read}
}
\value{
Returns a tibble with the read read replies.
}
\description{
Read the comment deletions
}
