% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{readCommentPart}
\alias{readCommentPart}
\title{Read the comment part of a comment}
\usage{
readCommentPart(path, files = NULL)
}
\arguments{
\item{path}{Path the the folder to read from}

\item{files}{The files to read, if NULL all files in the directory is read}
}
\value{
A tibble with the comments
}
\description{
Read the comment part of a comment
}
