% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{readComments}
\alias{readComments}
\title{Read and assembe the comment data}
\usage{
readComments(comment_setup, remove_duplicates = TRUE,
  read_files = checkCommentFolders(comment_setup))
}
\arguments{
\item{comment_setup}{comment_setup An object from
\code{\link{setUpCommentFolders}}}

\item{remove_duplicates}{Should duplicates be removed. This is especially for
edited comments where only the newest are kept in data if TRUE other wise all edits are kept}

\item{read_files}{The files to read.}
}
\value{
The current comment status
}
\description{
Read and assembe the comment data
}
