% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{readEdit}
\alias{readEdit}
\title{Read the comment part of an edited comment}
\usage{
readEdit(path, files = NULL)
}
\arguments{
\item{path}{Path the the folder to read from}

\item{files}{The files to read, if NULL all files in the directory is read}
}
\value{
A tibble with the edits
}
\description{
Read the comment part of an edited comment
}
