% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{readReply}
\alias{readReply}
\title{Read the replies}
\usage{
readReply(path, files = NULL)
}
\arguments{
\item{path}{path to the replies}

\item{files}{The files to read, if NULL all files are read}
}
\value{
Returns a tibble with the read read replies.
}
\description{
Read the replies
}
