% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{readStatus}
\alias{readStatus}
\title{Read status updates to comments}
\usage{
readStatus(path, files = NULL)
}
\arguments{
\item{path}{The path to the status updates}

\item{files}{the files t read. If NULL all files are read.}
}
\value{
Tibble with the status
}
\description{
Read status updates to comments
}
