% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{reviewProfile}
\alias{reviewProfile}
\title{set review status for a profile}
\usage{
reviewProfile(comment_data, user, profile)
}
\arguments{
\item{comment_data}{The comment data created by \cite{initCommentData}}

\item{user}{character with the initials of the user}

\item{profile}{character with the name of the profile}
}
\value{
Returns the updated comment_data
}
\description{
set review status for a profile
}
