% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{reviewStatus}
\alias{reviewStatus}
\title{Get the review status}
\usage{
reviewStatus(comment_data)
}
\arguments{
\item{comment_data}{The comment_data to update with review status}
}
\value{
returns a list with the users progress
}
\description{
Get the review status
}
