% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailfit.R
\name{tailFitRsquare.fit}
\alias{tailFitRsquare.fit}
\title{Get R2 from a fitted model}
\usage{
tailFitRsquare.fit(fit)
}
\arguments{
\item{fit}{A fitted linear model}
}
\value{
R-squared for the fitted model
}
\description{
Get R2 from a fitted model
}
\examples{
# Establish the data
atpt1n <- c(5760, 7200, 8640, 10080, 14400, 20160, 30360, 40320, 50400)
aval   <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 1500)
lower  <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 0)
upper  <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 3000)

# Fit the tail
tail_fit <- tailFit(aval, atpt1n)
tailFitRsquare.fit(tail_fit)
}
