% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailfit.R
\name{tHalf}
\alias{tHalf}
\title{Calculate t-half based on a fit}
\usage{
tHalf(fit)
}
\arguments{
\item{fit}{A tail fit based on either observed data or censored data}
}
\value{
t-half based on a tail fit
}
\description{
Calculate t-half based on a fit
}
\examples{
# Establish the data
atpt1n <- c(5760, 7200, 8640, 10080, 14400, 20160, 30360, 40320, 50400)
aval   <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 1500)
lower  <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 0)
upper  <- c(59630, 49690, 50130, 45410, 29980, 19650, 7488, 3332, 3000)

# Fit the tail
tail_fit <- tailFit(lower, atpt1n)
tHalf(tail_fit)

best_points <- bestTailFit(lower, atpt1n)$points
best_tail_fit <- tailFit(lower[best_points], atpt1n[best_points])

tHalf(best_tail_fit)

# Fit the tail based on the censored data
tail_fit_cens <- tailFitCens(lower, upper, atpt1n)
tHalf(tail_fit_cens)
}
