% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{updateCommentData}
\alias{updateCommentData}
\title{Update comment data according to current folder status}
\usage{
updateCommentData(comment_data)
}
\arguments{
\item{comment_data}{The comment data to update}
}
\value{
The updated comment data
}
\description{
Update comment data according to current folder status
}
