% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{updateCommentDataCheck}
\alias{updateCommentDataCheck}
\title{Check for files not currently in comment data}
\usage{
updateCommentDataCheck(comment_data,
  current_files = checkCommentFolders(attributes(comment_data)$comment_setup))
}
\arguments{
\item{comment_data}{The comment data object to check for updates}

\item{current_files}{optional, list of the the files currently loaded}
}
\value{
Return a list of the folders
}
\description{
Check for files not currently in comment data
}
