% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_rw_plotter.R
\name{wordplot}
\alias{wordplot}
\title{Title}
\usage{
wordplot(comb_plots, comments = NULL, type = "PK",
  title = c("NN1436-4225", "First draft", "PK-profiles",
  "Not validated"), ...)
}
\arguments{
\item{comb_plots}{list of plots to combine into a word document}

\item{comments}{data set with comments, created using createPlotComment}

\item{type}{The type of document, PK/PD}

\item{title}{The four titles for the header}

\item{...}{additional arguments passed on to header}
}
\value{
create the word file with the plots
}
\description{
Title
}
