% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{writeComment}
\alias{writeComment}
\title{Write a comment}
\usage{
writeComment(comment_data, user, profile, regarding, comment,
  time_points = numeric(0), view_range, paramcd = "", color_by = "",
  time_type = "point", time_point_paramcd = "")
}
\arguments{
\item{comment_data}{an object initialised by \code{\link{initCommentData}}.}

\item{user}{The initials of the user that gives the comment}

\item{profile}{The profiles which is associated with the comment}

\item{regarding}{What does the comment concern}

\item{comment}{What is the comment}

\item{time_points}{The time points which are associated with the comment}

\item{view_range}{The current x-axis range}

\item{paramcd}{The paramcd associated with the comment}

\item{color_by}{The color_by associated with the comment}

\item{time_type}{The type of points given to time_points. Can be "interval" or "point"}

\item{time_point_paramcd}{The paramcd associated with each time_point}
}
\value{
Adds the comment to the comment data object.
}
\description{
Write a comment
}
