% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{writeEdit}
\alias{writeEdit}
\title{Edit a comment}
\usage{
writeEdit(comment_data, comment_id, comment = NULL, regarding = NULL,
  paramcd = NULL, time_points = NULL, view_range = NULL)
}
\arguments{
\item{comment_data}{An object initialised by \code{\link{initCommentData}}.}

\item{comment_id}{The id of the comment to change.}

\item{comment}{The new comment, if NULL the comment text is not changed.}

\item{regarding}{The updated regarding}

\item{paramcd}{The updated paramcd}

\item{time_points}{The new time points, if NULL the time points are not updated.}

\item{view_range}{The updated view_range, if NULL the view_range is not updated.}
}
\value{
The edited comment data
}
\description{
Edit a comment
}
