% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{writeReply}
\alias{writeReply}
\title{Write a reply to a comment}
\usage{
writeReply(comment_data, comment_id, user, reply)
}
\arguments{
\item{comment_data}{An object initialised by \code{\link{initCommentData}}.}

\item{comment_id}{The id of the comment to change.}

\item{user}{The initials of the user replying.}

\item{reply}{Character with the reply.}
}
\value{
returns a tibble with the updated comment
}
\description{
Write a reply to a comment
}
