% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{writeStatus}
\alias{writeStatus}
\title{Update the status of a comment}
\usage{
writeStatus(comment_data, comment_id, user, status)
}
\arguments{
\item{comment_data}{An object initialised by \code{\link{initCommentData}}.}

\item{comment_id}{The id of the comment to change.}

\item{user}{The initials of the user replying.}

\item{status}{Character with the status.}
}
\value{
A tibble with the updated status.
}
\description{
Update the status of a comment
}
